Implemented rating system and functionality to ping officers.
The bot is ready for integration with nltk functions. When a message is sent, the message is stored in a variable which can then be used as an argument within a function that uses nltk to parse and return an output to the user. If the response is null, in this case - if a user types "Freeze!", then the bot pings a specific role which may be able to help them with their question. The ratings are stored in a .txt file with data separated by commas in the following order: question, response, user, and rating.